Added Scripts for Tag Synchronization and Teleportation
Teleportation script can teleport a list of objects to a new location in a way they will not overlap, then send a follow-up message to the objects to notify them of their teleportation.
Tag synchronization script synchronizes tags across the Photon network.  Please do not use this script if tags specify whether an object is locally owned (it is recommended to use Photon's built-in attributes to determine this instead).